chore: move testotel into internal
