+++ tests/extproc/vcr/otel_test.go
@@ -14,7 +14,7 @@ import (
 	"github.com/stretchr/testify/require"
 
 	"github.com/envoyproxy/ai-gateway/tests/internal/testenvironment"
-	"github.com/envoyproxy/ai-gateway/tests/testotel"
+	"github.com/envoyproxy/ai-gateway/tests/internal/testotel"
 )
 
 // otelTestEnvironment holds all the services needed for OTEL tests.
